requirements.txt
python-telegram-bot==8.0
requests==2.18.4
urllib3==1.22
